Add versions to the Analyzers.Version.targets Delete helpers and depend on Analyzer.Utitlies. Moved a couple of methods to the utilities project. Rename the test dll to follow the convention with the rest of the analyzers. Make the tests run on BuildAndTest.proj
