+++ PortFxCop/BuildAndTest.proj
@@ -63,7 +63,7 @@
       <TestAssemblies Include="Binaries\$(Configuration)\Microsoft.Composition.Analyzers.UnitTests.dll" />
       <TestAssemblies Include="Binaries\$(Configuration)\XmlDocumentationComments.Analyzers.UnitTests.dll" />
       <TestAssemblies Include="Binaries\$(Configuration)\System.Collections.Immutable.Analyzers.UnitTests.dll" />
-
+      <TestAssemblies Include="Binaries\$(Configuration)\Microsoft.CodeAnalysis.Analyzers.UnitTests.dll" />
     </ItemGroup>
 
     <Exec Command="&quot;$(PackagesDirectory)\xunit.runner.console.2.0.0\tools\xunit.console.x86.exe&quot; @(TestAssemblies, ' ') $(RunTestArgs)" />
