--- PortFxCop/src/Microsoft.CodeAnalysis.Analyzers/Core/Helpers/AttributeHelpers.cs
@@ -1,38 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System.Collections.Generic;
-
-namespace Microsoft.CodeAnalysis.Analyzers
-{
-    internal static class AttributeHelpers
-    {
-        internal static IEnumerable<AttributeData> GetApplicableAttributes(INamedTypeSymbol type)
-        {
-            var attributes = new List<AttributeData>();
-
-            while (type != null)
-            {
-                attributes.AddRange(type.GetAttributes());
-
-                type = type.BaseType;
-            }
-
-            return attributes;
-        }
-
-        internal static bool DerivesFrom(this INamedTypeSymbol symbol, INamedTypeSymbol candidateBaseType)
-        {
-            while (symbol != null)
-            {
-                if (symbol.Equals(candidateBaseType))
-                {
-                    return true;
-                }
-
-                symbol = symbol.BaseType;
-            }
-
-            return false;
-        }
-    }
-}
