--- PortFxCop/src/Microsoft.CodeAnalysis.Analyzers/Core/Helpers/ITypeSymbolExtensions.cs
@@ -1,29 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System.Collections.Generic;
-
-namespace Microsoft.CodeAnalysis.Analyzers
-{
-    internal static class ITypeSymbolExtensions
-    {
-        public static IEnumerable<ITypeSymbol> GetBaseTypesAndThis(this ITypeSymbol type)
-        {
-            var current = type;
-            while (current != null)
-            {
-                yield return current;
-                current = current.BaseType;
-            }
-        }
-
-        public static IEnumerable<INamedTypeSymbol> GetBaseTypes(this ITypeSymbol type)
-        {
-            var current = type.BaseType;
-            while (current != null)
-            {
-                yield return current;
-                current = current.BaseType;
-            }
-        }
-    }
-}
