+++ PortFxCop/src/Microsoft.CodeAnalysis.Analyzers/Core/InternalImplementationOnlyAnalyzer.cs
@@ -2,6 +2,7 @@
 
 using System.Collections.Immutable;
 using System.Linq;
+using Analyzer.Utilities;
 using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Microsoft.CodeAnalysis.Analyzers
@@ -19,7 +20,7 @@ namespace Microsoft.CodeAnalysis.Analyzers
                                                         DiagnosticIds.InternalImplementationOnlyRuleId,
                                                         s_localizableTitle,
                                                         s_localizableMessageFormat,
-                                                        DiagnosticCategory.Compatibility,
+                                                        AnalyzerDiagnosticCategory.Compatibility,
                                                         DiagnosticSeverity.Error,
                                                         true,
                                                         s_localizableDescription);
@@ -36,7 +37,7 @@ namespace Microsoft.CodeAnalysis.Analyzers
             // in "internals visible" context, then issue an error.
             foreach (var iface in namedTypeSymbol.AllInterfaces)
             {
-                var attributes = AttributeHelpers.GetApplicableAttributes(iface);
+                var attributes = iface.GetApplicableAttributes();
 
                 // We are doing a string comparison of the name here because we don't care where the attribute comes from.
                 // CodeAnalysis.dll itself has this attribute and if the user assembly also had it, symbol equality will fail
