+++ PortFxCop/src/Microsoft.CodeAnalysis.Analyzers/Core/MetaAnalyzers/RegisterActionAnalyzer.cs
@@ -25,7 +25,7 @@ namespace Microsoft.CodeAnalysis.Analyzers.MetaAnalyzers
             DiagnosticIds.MissingKindArgumentToRegisterActionRuleId,
             s_localizableTitleMissingKindArgument,
             s_localizableMessageMissingKindArgument,
-            DiagnosticCategory.AnalyzerCorrectness,
+            AnalyzerDiagnosticCategory.AnalyzerCorrectness,
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true,
             description: s_localizableDescriptionMissingKindArgument,
@@ -38,7 +38,7 @@ namespace Microsoft.CodeAnalysis.Analyzers.MetaAnalyzers
             DiagnosticIds.UnsupportedSymbolKindArgumentRuleId,
             s_localizableTitleUnsupportedSymbolKindArgument,
             s_localizableMessageUnsupportedSymbolKindArgument,
-            DiagnosticCategory.AnalyzerCorrectness,
+            AnalyzerDiagnosticCategory.AnalyzerCorrectness,
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true,
             customTags: WellKnownDiagnosticTags.Telemetry);
@@ -51,7 +51,7 @@ namespace Microsoft.CodeAnalysis.Analyzers.MetaAnalyzers
             DiagnosticIds.InvalidSyntaxKindTypeArgumentRuleId,
             s_localizableTitleInvalidSyntaxKindTypeArgument,
             s_localizableMessageInvalidSyntaxKindTypeArgument,
-            DiagnosticCategory.AnalyzerCorrectness,
+            AnalyzerDiagnosticCategory.AnalyzerCorrectness,
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true,
             description: s_localizableDescriptionInvalidSyntaxKindTypeArgument,
@@ -66,7 +66,7 @@ namespace Microsoft.CodeAnalysis.Analyzers.MetaAnalyzers
             DiagnosticIds.StartActionWithNoRegisteredActionsRuleId,
             s_localizableTitleStartActionWithNoRegisteredActions,
             s_localizableMessageStartActionWithNoRegisteredActions,
-            DiagnosticCategory.AnalyzerPerformance,
+            AnalyzerDiagnosticCategory.AnalyzerPerformance,
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true,
             description: s_localizableDescriptionStatefulAnalyzerRegisterActionsDescription,
@@ -79,7 +79,7 @@ namespace Microsoft.CodeAnalysis.Analyzers.MetaAnalyzers
             DiagnosticIds.StartActionWithOnlyEndActionRuleId,
             s_localizableTitleStartActionWithOnlyEndAction,
             s_localizableMessageStartActionWithOnlyEndAction,
-            DiagnosticCategory.AnalyzerPerformance,
+            AnalyzerDiagnosticCategory.AnalyzerPerformance,
             DiagnosticSeverity.Warning,
             isEnabledByDefault: true,
             description: s_localizableDescriptionStatefulAnalyzerRegisterActionsDescription,
